pulseaudio: Install a i.MX6ULL specific daemon.conf to increase the rlimit-rttime
The performance of imx6ull seems not good enough, it cause the pulseaudio exit
when cpu loading very high. The rlimit-rttime specifies a limit on the amount
of CPU time that a process scheduled under a real-time scheduling policy may
consume without making a blocking system call.  If the pulseaudio cost more
time which exceed the rlimit-rttime, it will be killed by kernel.
So here we increase the rlimit-rttime to workaround this issue.

Upstream-Status: Inappropriate [i.MX6ULL Specific]

Signed-off-by: Yuqing Zhu <carol.zhu@nxp.com>