systemd: Add work around for Weston' suspend/resume [YOCIMX-2095]
Weston got DRM permission failure with DRM operations during suspend and resume.
It conflict with the atomic modesetting patches for the weston 3.0.

The atomic modesetting patch split repaint into two stages, as implied by the grouped-repaint
interface: drm_output_repaint generates the repaint state only, and drm_repaint_flush
applies it. the new asynchronous rendering mode conflict with the systemd login.

This is a workaround patch, we will remove it after the conflict was fixed in the weston code.

Git for weston with atomic modesetting:
https://gitlab.collabora.com/daniels/weston/tree/wip/2017-12/atomic-v14

Signed-off-by: Yong Gan <yong.gan@nxp.com>
Signed-off-by: Jun Zhu <junzhu@nxp.com>